Added support for annotations identifying foreign code:
-["JavaExternClass": "java.util.List<Response<Void>>"]
-["JavaStaticClass": "java..lang.System"]

When the class is encountered after the annotation is is translated to
the String inside the extern annotation.

When a static class is encoutered a dummy object is created to use the
methods of the class. Upon new the object is registered and every future
encouter is translated to the Java static class.

!!! THESR ARE NOW REGISTRATIONS AS ABSTRACT TYPES IN THE JAVA
TRANSLATOR, WHICH I'M NOT 100% IS RIGHT